separated favouritelist component to sub components
